Initial Upload
This script was created to aid in migrations from Mac servers to
Windows servers. The script scans directories and files for characters
that are illegal in Windows environments. This particular script is not
perfect and has a few edge cases where it will not find ALL of the
possible combinations of characters and was really intended as more of
a way to assess a particular share as to whether it contained illegal
characters.